Adding support for headers in as part of Flutter Web a11y features using aria role "heading".
In addition to “Heading” role, in the web,  aria-level attribute is also used. This attribute which provides information of levels of headers compared to one another. Default header level is 2. In Flutter on the other hand the level information of headers is not provided.

The default level is not changed for now. We can add it later when Flutter provides that info.

TESTED: Unit tests + Manual tests in IOS (Chrome, Safari), Android (Chrome)
PiperOrigin-RevId: 254450974